refactor: [Util.kt] 예외 처리 함수를 Util 패키지로 분리
